Start refactoring features module
The aim here is to begin to remove functions that look like they will never be used and instead create functions that will return vectors of features for each measurement interval.  These can then be stored in a dictionary mapping user ids and activities to feature vectors and passed to classifiers.